+++ test/mocks/MockTradeService.sol
@@ -6,19 +6,17 @@ import { ITradeService } from "../../src/services/interfaces/ITradeService.sol";
 contract MockTradeService is ITradeService {
   struct IncreasePositionInputs {
     address _primaryAccount;
-    uint256 _subAccountId;
+    uint8 _subAccountId;
     uint256 _marketIndex;
     int256 _sizeDelta;
     uint256 _limitPriceE30;
   }
 
-
   struct DecreasePositionInputs {
     address _account;
-    uint256 _subAccountId;
+    uint8 _subAccountId;
     uint256 _marketIndex;
     uint256 _positionSizeE30ToDecrease;
-
     // @todo - support take profit token
     // address _tpToken;
     uint256 _limitPriceE30;
@@ -42,7 +40,7 @@ contract MockTradeService is ITradeService {
 
   function increasePosition(
     address _primaryAccount,
-    uint256 _subAccountId,
+    uint8 _subAccountId,
     uint256 _marketIndex,
     int256 _sizeDelta,
     uint256 _limitPriceE30
@@ -61,7 +59,7 @@ contract MockTradeService is ITradeService {
 
   function decreasePosition(
     address _account,
-    uint256 _subAccountId,
+    uint8 _subAccountId,
     uint256 _marketIndex,
     uint256 _positionSizeE30ToDecrease,
     address _tpToken,
@@ -74,17 +72,12 @@ contract MockTradeService is ITradeService {
         _subAccountId: _subAccountId,
         _marketIndex: _marketIndex,
         _positionSizeE30ToDecrease: _positionSizeE30ToDecrease,
-         _limitPriceE30: _limitPriceE30
+        _limitPriceE30: _limitPriceE30
       })
     );
   }
 
-  function forceClosePosition(
-    address _account,
-    uint256 _subAccountId,
-    uint256 _marketIndex,
-    address _tpToken
-  ) external {
+  function forceClosePosition(address _account, uint8 _subAccountId, uint256 _marketIndex, address _tpToken) external {
     decreasePositionCallCount++;
   }
 }
