+++ test/services/crossMargin/CrossMarginService_Base.t.sol
@@ -63,7 +63,7 @@ contract CrossMarginService_Base is BaseTest {
     vm.stopPrank();
   }
 
-  function getSubAccount(address _primary, uint256 _subAccountId) internal pure returns (address _subAccount) {
+  function getSubAccount(address _primary, uint8 _subAccountId) internal pure returns (address _subAccount) {
     if (_subAccountId > 255) revert();
     return address(uint160(_primary) ^ uint160(_subAccountId));
   }
